pyproject.toml
[project]
name = "hikingplots"
version = "0.1.0"
description = "A library to plot hikes (GPX files) onto maps"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "diskcache>=5.6.3",
    "earthpy>=0.9.4",
    "geopandas>=1.0.1",
    "geopy>=2.4.1",
    "gpxpy>=1.6.2",
    "lxml>=5.3.1",
    "matplotlib>=3.10.1",
    "overpy>=0.7",
    "pillow>=11.1.0",
    "pyyaml>=6.0.2",
    "rasterio>=1.4.3",
    "setuptools>=76.0.0",
    "shapely>=2.0.7",
    "termcolor>=2.5.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "ruff>=0.11.0",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "I", "F"]

[build-system]
requires = ["setuptools"]

[tool.setuptools.packages.find]
where = ["."]
include = ["hikingplots*"]
